requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.13.0
numpy==2.2.0
pillow==10.0.1
urllib3==2.0.7

# Web scraping and data processing
lxml==4.9.3
html5lib==1.1
soupsieve==2.5

# Visualization
matplotlib==3.8.0

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
pytz==2023.3

# File handling
fpdf==1.7.2
reportlab==4.0.4 